requirements.txt
Flask==3.0.0
requests==2.31.0
python-whois==0.8.0
dnspython==2.4.2
Werkzeug==3.0.0
python-dotenv==1.0.0
supabase>=2.4.0
websockets>=13.0
